requirements.txt
fastapi
uvicorn
websockets
google-genai
python-multipart
jinja2
